Fix dtype to fp16 in entropy variance script

+++ entropy_experiments/other_scripts/entropy_variance/entropy_variance.py
@@ -95,7 +95,7 @@ def load_model_and_tokenizer(backbone: str, adapter_path: str):
         base_id=backbone,
         adapter_path=adapter_path,
         mode="lora_simple",
-        dtype="fp32",  # Use fp32 to avoid BFloat16 issues
+        dtype="fp16",  # Use fp16 to avoid BFloat16 issues
         device_map="cuda",
         use_checkpointing=False,
     )
